+++ packages/shim/src/fs/virtual-files.js
@@ -1,8 +1,6 @@
 // Virtual plugin source served from memory; the fs shim's read path checks here before disk.
 
-function normalize(p) {
-  return (p || "").replace(/\\/g, "/").replace(/^\/+/, "").replace(/\/+$/, "");
-}
+import { normalize } from "../util/path.js";
 
 const virtualFiles = new Map();
